Moved Channel and MessageType from internal package
