+++ src/main/java/com/vonage/client/messages/messenger/MessengerFileRequest.java
@@ -18,7 +18,7 @@ package com.vonage.client.messages.messenger;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.vonage.client.messages.internal.MessagePayload;
-import com.vonage.client.messages.internal.MessageType;
+import com.vonage.client.messages.MessageType;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 public final class MessengerFileRequest extends MessengerRequest {
