+++ src/main/java/com/vonage/client/messages/messenger/MessengerRequest.java
@@ -18,7 +18,7 @@ package com.vonage.client.messages.messenger;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.vonage.client.messages.MessageRequest;
-import com.vonage.client.messages.internal.Channel;
+import com.vonage.client.messages.Channel;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 public abstract class MessengerRequest extends MessageRequest {
