+++ src/main/java/com/vonage/client/messages/viber/ViberRequest.java
@@ -18,7 +18,7 @@ package com.vonage.client.messages.viber;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.vonage.client.messages.MessageRequest;
-import com.vonage.client.messages.internal.Channel;
+import com.vonage.client.messages.Channel;
 import com.vonage.client.messages.internal.E164;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
