+++ src/main/java/com/vonage/client/messages/whatsapp/WhatsappTemplateRequest.java
@@ -17,7 +17,7 @@ package com.vonage.client.messages.whatsapp;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.vonage.client.messages.internal.MessageType;
+import com.vonage.client.messages.MessageType;
 
 import java.util.List;
 import java.util.Map;
