+++ src/test/java/com/vonage/client/messages/MessageRequestTest.java
@@ -15,13 +15,10 @@
  */
 package com.vonage.client.messages;
 
-import com.vonage.client.messages.internal.Channel;
-import com.vonage.client.messages.internal.MessageType;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 public class MessageRequestTest {
